+++ CHANGES.md
@@ -14,6 +14,7 @@
 - Removed terrible isinstance check of unittest.Mock in mirror.py - `PR #859` - Thanks **ichard26**
 - Put potential time consuming IO operations into executor - `PR #877`
 - Migrated Markdown documentation from recommonmark to MyST-Parser + docs config clean up - `PR #879` - Thanks **ichard26**
+- Use `shutil.move()` for temp file management - `PR #883`
 
 # 4.4.0 (2020-12-31)